Add files via upload
The analysis is done by creating various charts with the pivot table function in Excel.

The Dashboard gives insights on

No. of orders received and the sales done by the store.
The percentage of male and female customers.
Different sales channels of the store and the percentage of orders received from them.
Top states where the sales percentage is higher.
Analysis of the age & gender of customers.
And the status of the orders placed by the customers.